possible shift of analysis reporting config variables to ghsci
